Reorganize downscale into multiple files
